Added new entity.
Added GuardianKidPendingRelation.

Code refactor in order to extend from a new base class (GuardianKidRelationBase).
Repository code refactor in order to have a common base repo (GuardianKidRelationBaseRepository).
