+++ src/Entity/GuardianKidPendingRelation.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Entity;
+
+use App\Repository\GuardianKidPendingRelationRepository;
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity(repositoryClass=GuardianKidPendingRelationRepository::class)
+ * @ORM\Table(name="guardian_kid_pending_relation")
+ *
+ * Class GuardianKidPendingRelation
+ * @package App\Entity
+ */
+class GuardianKidPendingRelation extends GuardianKidRelationBase
+{
+}
