+++ src/Repository/GuardianKidRelationBaseRepository.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Repository;
+
+abstract class GuardianKidRelationBaseRepository extends BaseRepository
+{
+}
